requirements.txt
python-dotenv==1.1.0
Flask==3.1.1
SQLAlchemy==2.0.40
Flask-SQLAlchemy==3.1.1
flask-restx==1.3.0
flask-cors==5.0.1
Flask-JWT-Extended==4.7.1
psycopg2==2.9.10
requests==2.32.3
redis==6.2.0